Update NeoFaker.Text module
